navigation bar fix + activities to fragments
