+++ app/src/main/java/com/example/tindercut/DataSerializable.java
@@ -5,7 +5,7 @@ import java.io.Serializable;
 public class DataSerializable implements Serializable {
     String Data;
 
-    DataSerializable(String Data) {
+    public DataSerializable(String Data) {
         super();
         this.Data = Data;
     }
