--- app/src/main/java/com/example/tindercut/HomeActivity.java
@@ -1,46 +1,0 @@
-package com.example.tindercut;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.MenuItem;
-import android.view.View;
-
-import com.google.android.material.bottomnavigation.BottomNavigationView;
-
-public class HomeActivity extends AppCompatActivity {
-
-    private BottomNavigationView bottomNavigationView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_home);
-
-        bottomNavigationView = findViewById(R.id.bottom_navigator);
-        bottomNavigationView.setSelectedItemId(R.id.navigation_home);
-
-        bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                switch (item.getItemId()){
-                    case R.id.navigation_searchphoto:
-                        startActivity(new Intent(getApplicationContext(), MainActivity.class));
-                        overridePendingTransition(0, 0);
-                        return true;
-                    case R.id.navigation_profile:
-                        startActivity(new Intent(getApplicationContext(), HairdresserActivity.class));
-                        overridePendingTransition(0, 0);
-                        return true;
-                    case R.id.navigation_home:
-                        return true;
-                }
-                return false;
-            }
-
-        });
-
-    }
-}
