+++ app/src/main/java/com/example/tindercut/MainActivity.java
@@ -1,241 +1,84 @@
 package com.example.tindercut;
 
-import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-import android.Manifest;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
-import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
-import android.os.Environment;
-import android.provider.MediaStore;
-import android.provider.Settings;
-import android.util.Base64;
-import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
-import android.widget.Toast;
 
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-import androidx.core.content.ContextCompat;
-
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.error.VolleyError;
-import com.android.volley.request.SimpleMultiPartRequest;
-import com.android.volley.toolbox.Volley;
+import com.example.tindercut.data.User;
+import com.example.tindercut.ui.hairdresser.HairdresserFragment;
+import com.example.tindercut.ui.home.HomeFragment;
+import com.example.tindercut.ui.login.LoginActivity;
+import com.example.tindercut.ui.photosearch.SearchPhotoFragment;
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationBarView;
-import com.hbisoft.pickit.PickiT;
-import com.hbisoft.pickit.PickiTCallbacks;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Calendar;
-
-public class MainActivity extends AppCompatActivity implements PickiTCallbacks {
-
-    private final int picked_image = 1;
-    PickiT pickIt;
-    private Bitmap ImageBitmap;
-    private Button button;
-    private Button button2;
-    private ImageView ImageView;
-    private String imagePath;
+public class MainActivity extends AppCompatActivity {
 
     private BottomNavigationView bottomNavigationView;
+    private final Fragment fragment1 = new SearchPhotoFragment();
+    private final Fragment fragment2 = new HomeFragment();
+    private final Fragment fragment3 = new HairdresserFragment();
+    private final FragmentManager fm = getSupportFragmentManager();
+    private Fragment active;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
+        fm.beginTransaction().add(R.id.main_container, fragment3, "3").hide(fragment3).commit();
+        fm.beginTransaction().add(R.id.main_container, fragment2, "2").hide(fragment2).commit();
+        fm.beginTransaction().add(R.id.main_container ,fragment1, "1").commit();
 
-        pickIt = new PickiT(this, this, this);
+        active = fragment1;
 
-        //Привязка обработчика события к кнопке
-        button = findViewById(R.id.button);
-        button.setOnClickListener(new View.OnClickListener() {
+        bottomNavigationView = findViewById(R.id.bottom_navigator);
+        bottomNavigationView.setSelectedItemId(R.id.navigation_searchphoto);
+
+        bottomNavigationView.setOnNavigationItemSelectedListener(new BottomNavigationView.OnNavigationItemSelectedListener() {
             @Override
-            public void onClick(View v) {
-                Intent pickPhoto = new Intent(Intent.ACTION_PICK);
-                pickPhoto.setType("image/*");
-                pickPhoto.setAction(Intent.ACTION_GET_CONTENT);
-                checkPermission();
-                startActivityForResult(pickPhoto, picked_image);
-            }
-        });
-
-        button2 = findViewById(R.id.button2);
-        button2.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (ImageBitmap != null) {
-                    sendImage();
-                }
-            }
-        });
-
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        switch (requestCode) {
-            case picked_image:
-                //Получаем URI картинки:
-                if (data == null) {
-                    return;
-                }
-                Uri imageUri = data.getData();
-                pickIt.getPath(imageUri, Build.VERSION.SDK_INT);
-                //Находим нужный ImageView в интерфейсе
-                ImageView = findViewById(R.id.imageView);
-
-                try {
-                    ImageBitmap = MediaStore.Images.Media.getBitmap(getContentResolver(), imageUri);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-
-                ByteArrayOutputStream byteArrayOutputStream;
-                byteArrayOutputStream = new ByteArrayOutputStream();
-
-                if (ImageBitmap != null) {
-                    ImageBitmap.compress(Bitmap.CompressFormat.JPEG, 100, byteArrayOutputStream);
-                    byte[] bytes = byteArrayOutputStream.toByteArray();
-                    final String base64Image = Base64.encodeToString(bytes, Base64.DEFAULT);
-                    ImageView.setImageBitmap(ImageBitmap);
-                }
-
-        }
-
-
-    }
-
-    public void sendImage() {
-        RequestQueue queue = Volley.newRequestQueue(getApplicationContext());
-        String url = "http://79.137.206.63:8011/img";
-
-        SimpleMultiPartRequest uploadRequest = new SimpleMultiPartRequest(Request.Method.POST, url,
-                new Response.Listener<String>() {
-                    @Override
-                    public void onResponse(String response) {
-                        try {
-                            JSONObject responseJSON = new JSONObject(response);
-                            JSONArray hairData = responseJSON.getJSONArray("data");
-
-                            //Выполняем проверку успеха отправки изображения
-                            String requestResult = responseJSON.getString("result");
-                            if (requestResult.equals("Ok")) {
-                                Toast.makeText(getApplicationContext(), "Изображение успешно загружено.", Toast.LENGTH_SHORT).show();
-
-                                Intent scrollingIntent = new Intent(MainActivity.this, ScrollingActivity.class);
-                                ArrayList<DataSerializable> arrayToSend = new ArrayList<>();
-                                for (int i = 0; i < hairData.length(); i++) {
-                                    arrayToSend.add(new DataSerializable(hairData.get(i).toString()));
-                                }
-
-                                Bundle bundle = new Bundle();
-                                bundle.putSerializable("dataArray", arrayToSend);
-                                scrollingIntent.putExtras(bundle);
-                                try {
-                                    startActivity(scrollingIntent);
-                                } catch (Exception e) {
-                                    e.getStackTrace();
-                                }
-
-                            } else {
-                                Toast.makeText(getApplicationContext(), requestResult, Toast.LENGTH_SHORT).show();
-                            }
-
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                switch (item.getItemId()) {
+                    case R.id.navigation_searchphoto:
+                        fm.beginTransaction().hide(active).show(fragment1).commit();
+                        active = fragment1;
+                        return true;
+                    case R.id.navigation_profile:
+                        if (!User.isLogin(getApplicationContext()) && !LoggedInWithGoogleAuth()) {
+                            openLoginActivity();
                         }
-                    }
-                }, new Response.ErrorListener() {
-            @Override
-            public void onErrorResponse(VolleyError error) {
-                Toast.makeText(getApplicationContext(), error.getLocalizedMessage(), Toast.LENGTH_LONG).show();
+
+                        fm.beginTransaction().hide(active).show(fragment3).commit();
+                        active = fragment3;
+                        return true;
+                    case R.id.navigation_home:
+                        fm.beginTransaction().hide(active).show(fragment2).commit();
+                        active = fragment2;
+                        return true;
+                }
+                return false;
             }
         });
-        String fileName = Calendar.getInstance().getTime() + " Photo";
-        //System.out.println("Путь: "+ imageUri.getPath());
-        uploadRequest.addFile(fileName, imagePath);
-        //uploadRequest.addMultipartParam("body", "text/plain", base64Image);
-
-        queue.add(uploadRequest);
-    }
-
-    @Override
-    public void PickiTonUriReturned() {
 
     }
 
-    @Override
-    public void PickiTonStartListener() {
-
+    private void openLoginActivity() {
+        Intent intent = new Intent(getApplicationContext(), LoginActivity.class);
+        startActivity(intent);
+        finish();
     }
-
-    @Override
-    public void PickiTonProgressUpdate(int progress) {
-
-    }
-
-    @Override
-    public void PickiTonCompleteListener(String path, boolean wasDriveFile, boolean wasUnknownProvider, boolean wasSuccessful, String Reason) {
-
-        imagePath = path;
-
-    }
-
-    @Override
-    public void PickiTonMultipleCompleteListener(ArrayList<String> paths, boolean wasSuccessful, String Reason) {
-
-    }
-
-    public void checkPermission() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            if (ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
-                ActivityCompat.requestPermissions(MainActivity.this, new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, 100);
-
-            }
-
-            if (ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.ACCESS_NETWORK_STATE) != PackageManager.PERMISSION_GRANTED) {
-                ActivityCompat.requestPermissions(MainActivity.this, new String[]{Manifest.permission.ACCESS_NETWORK_STATE}, 101);
-
-            }
-        }
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            if (!Environment.isExternalStorageManager()) {
-                try {
-                    Intent intent = new Intent(Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION);
-                    intent.addCategory("android.intent.category.DEFAULT");
-                    intent.setData(Uri.parse(String.format("package:%s", getApplicationContext().getPackageName())));
-                } catch (Exception e) {
-                    Intent intent = new Intent();
-                    intent.setAction(Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION);
-
-                }
-            }
-        }
+    private Boolean LoggedInWithGoogleAuth() {
+        GoogleSignInAccount account = GoogleSignIn.getLastSignedInAccount(getApplicationContext());
+        return account != null;
     }
 }
