+++ app/src/main/java/com/example/tindercut/ui/home/HomeFragment.java
@@ -1,0 +1,28 @@
+package com.example.tindercut.ui.home;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.tindercut.R;
+
+public class HomeFragment extends Fragment {
+
+    TextView textView;
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_home,
+                container, false);
+
+    textView = view.findViewById(R.id.inDevText);
+    return view;
+
+    }
+
+}
