+++ app/src/main/java/com/example/tindercut/ui/navigation/NavigationFragment.java
@@ -1,18 +1,12 @@
 package com.example.tindercut.ui.navigation;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Button;
 
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
-import com.example.tindercut.HairdresserActivity;
-import com.example.tindercut.HomeActivity;
 import com.example.tindercut.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
